requirements.txt
asgiref==3.9.1
boto3==1.40.11
botocore==1.40.11
Django==5.2.5
django-rest-framework==0.1.0
djangorestframework==3.16.1
jmespath==1.0.1
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
